backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic>=2.0
pydantic-settings
email-validator
